requirements.txt
numpy>=1.26
scipy>=1.10
pandas>=2.0
matplotlib>=3.8

pyyaml
hydra-core>=1.3
omegaconf>=2.3

jax[cuda12]

e3x

ase>=3.22
torch==2.3.*   # GPU/CPU build will resolve automatically depending on env
torchani>=2.2

scikit-learn>=1.3
statsmodels>=0.14
threadpoolctl>=3.1
tabulate>=0.9
traitlets>=5.14
typeguard>=4.1
toml
tqdm>=4.66
toolz>=0.12
wheel>=0.41
zipp>=3.17

# CHARMM optional — install separately if needed
# pycharmm

